requirements.txt
a2wsgi==1.10.8
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiologic==0.14.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==2.2.4
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.2
apispec==3.3.2
argcomplete==2.1.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==20.3.0
babel==2.17.0
backports.strenum==1.3.1
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
cadwyn==5.3.3
cattrs==1.10.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==7.1.2
clickclick==20.10.2
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.2
cycler==0.12.1
databricks-sdk==0.50.0
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
dnspython==2.7.0
docker==7.1.0
docutils==0.16
dotenv==0.9.9
email-validator==1.3.1
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
Flask==1.1.4
Flask-AppBuilder==3.4.4
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==3.25.1
Flask-Limiter==3.12
Flask-Login==0.4.1
Flask-OpenID==1.3.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-re2==1.1.20240702
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.6.1
inflection==0.5.1
iso8601==2.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==3.2.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy-object-proxy==1.11.0
libcst==1.7.0
limits==5.0.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.26.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.10.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
mlflow==2.21.3
mlflow-skinny==2.21.3
more-itertools==10.7.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.3
networkx==3.4.2
numpy==2.2.5
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pendulum==2.1.2
pillow==11.2.1
pluggy==1.5.0
prison==0.2.1
propcache==0.3.1
protobuf==5.29.4
psutil==5.9.8
psycopg2==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==1.7.1
pyparsing==3.2.3
pyrsistent==0.20.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2025.2
pytzdata==2020.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
retryhttp==1.3.2
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rich-toolkit==0.14.1
rpds-py==0.24.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
setproctitle==1.3.5
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.46.2
structlog==25.2.0
svcs==25.1.0
swagger-ui-bundle==0.0.9
sympy==1.13.1
tabulate==0.8.10
tenacity==9.1.2
termcolor==3.0.1
text-unidecode==1.3
threadpoolctl==3.6.0
torch==2.6.0
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.6
urllib3==2.4.0
uuid6==2024.7.10
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==1.0.1
wirerope==1.0.0
wrapt==1.17.2
WTForms==2.3.3
yarl==1.20.0
zipp==3.21.0
